finish punishing someone

+++ LibraryApp.Data/Entity/User.cs
@@ -14,7 +14,7 @@ namespace LibraryApp.Data.Entity
         public DateTime BirthDate { get; set; }
         public char Gender { get; set; }
         public bool IsPunished { get; set; }
-        public int FineAmount { get; set; }
+        public float FineAmount { get; set; }
 
         public Role Role { get; set; } = null!;
         public List<BookAuthor> BookAuthors { get; set; } = new();
